+++ src/Schema/Directives/ModelMutationDirective.php
@@ -6,38 +6,16 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Nuwave\Lighthouse\Execution\Arguments\ArgumentSet;
 use Nuwave\Lighthouse\Execution\Arguments\ResolveNested;
-use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Execution\TransactionalMutations;
-use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\ArgResolver;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Nuwave\Lighthouse\Support\Utils;
 
-abstract class MutationExecutorDirective extends BaseDirective implements FieldResolver, ArgResolver
+abstract class ModelMutationDirective extends BaseDirective implements FieldResolver, ArgResolver
 {
     public function __construct(
         protected TransactionalMutations $transactionalMutations,
     ) {}
-
-    public function resolveField(FieldValue $fieldValue): callable
-    {
-        $modelClass = $this->getModelClass();
-
-        return function (mixed $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($modelClass): Model {
-            $model = new $modelClass();
-
-            return $this->transactionalMutations->execute(
-                function () use ($model, $resolveInfo): Model {
-                    $mutated = $this->executeMutation($model, $resolveInfo->argumentSet);
-                    assert($mutated instanceof Model);
-
-                    return $mutated->refresh();
-                },
-                $model->getConnectionName(),
-            );
-        };
-    }
 
     /**
      * @param  Model  $model